requirements.txt
Flask==3.0.2
PyJWT==2.9.0
joblib==1.5.3
numpy==1.26.4
supabase==2.4.3